requirements.txt
aiodataloader==0.4.0
amqp==5.2.0
aniso8601==9.0.1
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
bcrypt==4.1.2
billiard==4.2.0
celery==5.4.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudinary==1.40.0
cryptography==42.0.7
dj-rest-auth==6.0.0
Django==5.0.7
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==23.5
django-ratelimit==4.1.0
django-stubs==4.2.7
django-stubs-ext==5.0.2
djangorestframework==3.15.2
exceptiongroup==1.2.1
Faker==21.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
h11==0.14.0
idna==3.7
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.11.2
kombu==5.3.7
libcst==1.4.0
Markdown==3.5.1
markdown-it-py==3.0.0
mdurl==0.1.2
mypy==1.7.1
mypy-extensions==1.0.0
oauthlib==3.2.2
pillow==10.3.0
promise==2.3
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.22
Pygments==2.18.0
PyJWT==2.8.0
pypng==0.20220715.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
referencing==0.32.0
requests==2.32.2
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.15.2
Rx==1.6.3
shellingham==1.5.4
singledispatch==4.1.0
six==1.16.0
sniffio==1.3.1
sqlparse==0.4.4
starlette==0.37.2
text-unidecode==1.3
tomli==2.0.1
typer==0.12.3
types-pytz==2024.1.0.20240417
types-PyYAML==6.0.12.20240311
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.30.1
vine==5.1.0
wcwidth==0.2.13

shortuuid~=1.0.13